FIXUP: nearbyintf is fair enough

+++ Manam.h
@@ -69,9 +69,9 @@ constexpr int STUDIO_SEND = 5;
 constexpr int STUDIO_FILTR = 6;
 constexpr int STUDIO_MSG = 7;
 // gen eps
-constexpr double mm2pkt = 0.2835; // 1mm = 2.835pt
-constexpr double pkt2mm = 0.3527; // 1pt = 0.3527mm
-constexpr double pkt_10m = 2.835;
+constexpr float mm2pkt = 0.2835; // 1mm = 2.835pt
+constexpr float pkt2mm = 0.3527; // 1pt = 0.3527mm
+constexpr float pkt_10m = 2.835;
 constexpr int podpisH = 6; // 2.1162 [mm]
 constexpr int preview_offset = 30;
 // format materialu
